Change per oneAPI 2022.3 release notes
https://www.intel.com/content/www/us/en/developer/articles/release-notes/intel-oneapi-dpc-c-compiler-release-notes.html

CPU driver is now notified of the location of TBB dynamic libraries via entry in cl.cfg file
located in the package.

This commit removes opencl-win-activate scripts, and instead modifies cl.cfg so that

CL_CONFIG_TBB_DLL_PATH = %LIBRARY_BIN%

I have not found other way that to use Python to do that replacement, so Python was added as host dependency
for intel-opencl-rt repackaging step.